requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
PyMySQL